feat: enhance backend functionality with new routes, controllers, and unique constraints; refactor existing code for improved structure and error handling

+++ backend/src/controllers/cartController.js
@@ -5,7 +5,7 @@ export async function getUserCart(req, res) {
     const userId = Number(req.userId)
 
     try {
-        const cart = await prisma.cartItem.findFirst({
+        const cart = await prisma.cart.findFirst({
             where: { userId },
             select: basicCartInfo
         })
@@ -57,9 +57,25 @@ export async function addItemToCart(req, res) {
 }
 
 export async function deleteItemFromCart(req, res) {
-    const { id } = req.body
+    const id = Number(req.params.id)
+    const userId = Number(req.userId)
 
     try {
+        const itemToDelete = await prisma.cartItem.findFirst({
+            where: { id },
+            select: {
+                cart: {
+                    select: {
+                        userId: true
+                    }
+                }
+            }
+        })
+
+        if (!itemToDelete) return res.status(404).json({ message: "Item not found" })
+
+        if (itemToDelete.cart.userId !== userId) return res.status(403).json({ message: "Action unauthorized" })
+
         const deletedCartItem = await prisma.cartItem.delete({
             where: { id }
         })
@@ -71,10 +87,26 @@ export async function deleteItemFromCart(req, res) {
 }
 
 export async function decreaseCartItemQuantity(req, res) {
-    const { id } = req.body
+    const id = Number(req.params.id)
+    const userId = Number(req.userId)
     let updatedCartItem;
 
     try {
+        const itemToUpdate = await prisma.cartItem.findFirst({
+            where: { id },
+            select: {
+                cart: {
+                    select: {
+                        userId: true
+                    }
+                }
+            }
+        })
+
+        if (!itemToUpdate) return res.status(404).json({ message: "Item not found" })
+
+        if (itemToUpdate.cart.userId !== userId) return res.status(403).json({ message: "Action unauthorized" })
+        
         const cartItem = await prisma.cartItem.findUnique({
             where: { id }
         })
